optimize the statistic of downloading
