--- modules/messagequeen/infrastructure/kafka/impl.go
@@ -1,16 +1,0 @@
-package kafka
-
-import (
-	"encoding/json"
-
-	kfklib "github.com/opensourceways/kafka-lib/agent"
-)
-
-func Publish(topic string, v interface{}, header map[string]string) error {
-	body, err := json.Marshal(v)
-	if err != nil {
-		return err
-	}
-
-	return kfklib.Publish(topic, header, body)
-}
