+++ modules/messagequeue/mq.go
@@ -1,4 +1,4 @@
-package messagequeen
+package messagequeue
 
 import (
 	"code.gitea.io/gitea/modules/log"
@@ -8,5 +8,6 @@ import (
 // Init the message queen, (ex: ActiveMQ、RocketMQ、RabbitMQ、Kafka)
 func Init() (err error) {
 	log.Info("Initialising message queen with type: %s", setting.MQ.MessageType)
+
 	return newKafkaMessageQueue(setting.MQ)
 }
