+++ routers/web/githttp.go
@@ -4,13 +4,13 @@
 package web
 
 import (
-	"code.gitea.io/gitea/routers/web/statmiddleware"
 	"net/http"
 
 	"code.gitea.io/gitea/modules/context"
 	"code.gitea.io/gitea/modules/setting"
 	"code.gitea.io/gitea/modules/web"
 	"code.gitea.io/gitea/routers/web/repo"
+	"code.gitea.io/gitea/routers/web/statmiddleware"
 	context_service "code.gitea.io/gitea/services/context"
 )
 
@@ -30,7 +30,11 @@ func requireSignIn(ctx *context.Context) {
 func gitHTTPRouters(m *web.Route) {
 	m.Group("", func() {
 		m.Group("", func() {
-			m.PostOptions("/git-upload-pack", statmiddleware.GitCloneMessageMiddleware, repo.ServiceUploadPack)
+			m.PostOptions(
+				"/git-upload-pack",
+				statmiddleware.GitCloneMessageMiddleware,
+				repo.ServiceUploadPack,
+			)
 		}, context.RepoAssignment)
 
 		m.PostOptions("/git-receive-pack", repo.ServiceReceivePack)
